+++ CHANGELOG.md
@@ -1,2 +1,11 @@
-2/9/14
+#####7/12/14
+
+ * added a script to install [Karma](https://github.com/InformationIntegrationGroup/Web-Karma).
+
+#####7/3/14
+
+ * changed the default install to be VIVO 1.7.  The database name has also been changed to `vivo17dev`.  VIVO 1.6 and VIVO 1.5 are still available in separate branches.
+
+#####2/9/14
+
  * changed default install to be VIVO 1.6.  Use the v1.5 branch if you would like to install VIVO 1.5.